% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R, R/Subcollection.R
\name{add}
\alias{add}
\title{add}
\usage{
add(content)

add(content)
}
\arguments{
\item{content}{sth to be added}

\item{relativePath}{path to add sth}
}
\value{
Collection object.

Collection object.
}
\description{
Adds ArvadosFile or Subcollection specified by content to the collection.

Adds ArvadosFile or Subcollection specified by content to the subcollection.
}
