% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations.create}
\alias{api_client_authorizations.create}
\title{api_client_authorizations.create}
\usage{
arv$api_client_authorizations.create(apiclientauthorization,
	ensure_unique_name = "false")
}
\arguments{
\item{apiClientAuthorization}{ApiClientAuthorization object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations.create is a method defined in Arvados class.
}
