% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations.current}
\alias{api_client_authorizations.current}
\title{api_client_authorizations.current}
\usage{
arv$api_client_authorizations.current(NULL)
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations.current is a method defined in Arvados class.
}
