% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations.show}
\alias{api_client_authorizations.show}
\title{api_client_authorizations.show is a method defined in Arvados class.}
\usage{
arv$api_client_authorizations.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations.show is a method defined in Arvados class.
}
