% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations_create}
\alias{api_client_authorizations_create}
\title{api_client_authorizations_create}
\usage{
arv$api_client_authorizations_create(apiclientauthorization,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{apiClientAuthorization}{ApiClientAuthorization object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations_create is a method defined in Arvados class.
}
