% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations_delete}
\alias{api_client_authorizations_delete}
\title{api_client_authorizations_delete}
\usage{
arv$api_client_authorizations_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the ApiClientAuthorization in question.}
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations_delete is a method defined in Arvados class.
}
