% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_client_authorizations_update}
\alias{api_client_authorizations_update}
\title{api_client_authorizations_update}
\usage{
arv$api_client_authorizations_update(apiclientauthorization,
	uuid)
}
\arguments{
\item{apiClientAuthorization}{ApiClientAuthorization object.}

\item{uuid}{The UUID of the ApiClientAuthorization in question.}
}
\value{
ApiClientAuthorization object.
}
\description{
api_client_authorizations_update is a method defined in Arvados class.
}
