% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_clients.create}
\alias{api_clients.create}
\title{api_clients.create}
\usage{
arv$api_clients.create(apiclient,
	ensure_unique_name = "false")
}
\arguments{
\item{apiClient}{ApiClient object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
ApiClient object.
}
\description{
api_clients.create is a method defined in Arvados class.
}
