% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_clients.delete}
\alias{api_clients.delete}
\title{api_clients.delete}
\usage{
arv$api_clients.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the ApiClient in question.}
}
\value{
ApiClient object.
}
\description{
api_clients.delete is a method defined in Arvados class.
}
