% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_clients.show}
\alias{api_clients.show}
\title{api_clients.show is a method defined in Arvados class.}
\usage{
arv$api_clients.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
ApiClient object.
}
\description{
api_clients.show is a method defined in Arvados class.
}
