% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{api_clients.update}
\alias{api_clients.update}
\title{api_clients.update}
\usage{
arv$api_clients.update(apiclient,
	uuid)
}
\arguments{
\item{apiClient}{ApiClient object.}

\item{uuid}{The UUID of the ApiClient in question.}
}
\value{
ApiClient object.
}
\description{
api_clients.update is a method defined in Arvados class.
}
