% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient}
\description{
ApiClient
}
\section{Usage}{

\preformatted{apiClient -> ApiClient$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, name = NULL,
	url_prefix = NULL, created_at = NULL, updated_at = NULL,
	is_trusted = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{name}{}
    \item{url_prefix}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{is_trusted}{}
  }
}

