% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ApiClientAuthorization}
\alias{ApiClientAuthorization}
\title{ApiClientAuthorization}
\description{
ApiClientAuthorization
}
\section{Usage}{

\preformatted{apiClientAuthorization -> ApiClientAuthorization$new(uuid = NULL,
	etag = NULL, api_token = NULL, api_client_id = NULL,
	user_id = NULL, created_by_ip_address = NULL, last_used_by_ip_address = NULL,
	last_used_at = NULL, expires_at = NULL, created_at = NULL,
	updated_at = NULL, default_owner_uuid = NULL, scopes = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{api_token}{}
    \item{api_client_id}{}
    \item{user_id}{}
    \item{created_by_ip_address}{}
    \item{last_used_by_ip_address}{}
    \item{last_used_at}{}
    \item{expires_at}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{default_owner_uuid}{}
    \item{scopes}{}
    \item{uuid}{}
  }
}

