% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ApiClientAuthorizationList}
\alias{ApiClientAuthorizationList}
\title{ApiClientAuthorizationList}
\description{
ApiClientAuthorization list
}
\section{Usage}{

\preformatted{apiClientAuthorizationList -> ApiClientAuthorizationList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#apiClientAuthorizationList.}
    \item{etag}{List version.}
    \item{items}{The list of ApiClientAuthorizations.}
    \item{next_link}{A link to the next page of ApiClientAuthorizations.}
    \item{next_page_token}{The page token for the next page of ApiClientAuthorizations.}
    \item{selfLink}{A link back to this list.}
  }
}

