% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ApiClientList}
\alias{ApiClientList}
\title{ApiClientList}
\description{
ApiClient list
}
\section{Usage}{

\preformatted{apiClientList -> ApiClientList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#apiClientList.}
    \item{etag}{List version.}
    \item{items}{The list of ApiClients.}
    \item{next_link}{A link to the next page of ApiClients.}
    \item{next_page_token}{The page token for the next page of ApiClients.}
    \item{selfLink}{A link back to this list.}
  }
}

