% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\docType{class}
\name{Arvados-class}
\alias{Arvados-class}
\alias{Arvados}
\title{Arvados SDK Object}
\description{
All Arvados logic is inside this class
}
\section{Fields}{

\describe{
\item{\code{token}}{represents user authentification token.}

\item{\code{host}}{represents server name we wish to connect to.}
}}


