% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosFile.R
\docType{data}
\name{ArvadosFile}
\alias{ArvadosFile}
\title{ArvadosFile Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ArvadosFile
}
\description{
Update description
}
\keyword{datasets}
