% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{authorized_keys.create}
\alias{authorized_keys.create}
\title{authorized_keys.create}
\usage{
arv$authorized_keys.create(authorizedkey,
	ensure_unique_name = "false")
}
\arguments{
\item{authorizedKey}{AuthorizedKey object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
AuthorizedKey object.
}
\description{
authorized_keys.create is a method defined in Arvados class.
}
