% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{authorized_keys.destroy}
\alias{authorized_keys.destroy}
\title{authorized_keys.destroy is a method defined in Arvados class.}
\usage{
arv$authorized_keys.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
AuthorizedKey object.
}
\description{
authorized_keys.destroy is a method defined in Arvados class.
}
