% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{authorized_keys.get}
\alias{authorized_keys.get}
\title{authorized_keys.get}
\usage{
arv$authorized_keys.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the AuthorizedKey in question.}
}
\value{
AuthorizedKey object.
}
\description{
authorized_keys.get is a method defined in Arvados class.
}
