% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{authorized_keys_create}
\alias{authorized_keys_create}
\title{authorized_keys_create}
\usage{
arv$authorized_keys_create(authorizedkey,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{authorizedKey}{AuthorizedKey object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
AuthorizedKey object.
}
\description{
authorized_keys_create is a method defined in Arvados class.
}
