% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{AuthorizedKey}
\alias{AuthorizedKey}
\title{AuthorizedKey}
\description{
AuthorizedKey
}
\section{Usage}{

\preformatted{authorizedKey -> AuthorizedKey$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, name = NULL,
	key_type = NULL, authorized_user_uuid = NULL, public_key = NULL,
	expires_at = NULL, created_at = NULL, updated_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{name}{}
    \item{key_type}{}
    \item{authorized_user_uuid}{}
    \item{public_key}{}
    \item{expires_at}{}
    \item{created_at}{}
    \item{updated_at}{}
  }
}

