% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{AuthorizedKeyList}
\alias{AuthorizedKeyList}
\title{AuthorizedKeyList}
\description{
AuthorizedKey list
}
\section{Usage}{

\preformatted{authorizedKeyList -> AuthorizedKeyList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#authorizedKeyList.}
    \item{etag}{List version.}
    \item{items}{The list of AuthorizedKeys.}
    \item{next_link}{A link to the next page of AuthorizedKeys.}
    \item{next_page_token}{The page token for the next page of AuthorizedKeys.}
    \item{selfLink}{A link back to this list.}
  }
}

