% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\docType{class}
\name{Collection-class}
\alias{Collection-class}
\alias{Collection}
\title{Collection Object}
\arguments{
\item{uuid}{Object ID}

\item{etag}{Object version}

\item{owner_uuid}{No description}

\item{created_at}{No description}

\item{modified_by_client_uuid}{No description}

\item{modified_by_user_uuid}{No description}

\item{modified_at}{No description}

\item{portable_data_hash}{No description}

\item{replication_desired}{No description}

\item{replication_confirmed_at}{No description}

\item{replication_confirmed}{No description}

\item{updated_at}{No description}

\item{manifest_text}{No description}

\item{name}{No description}

\item{description}{No description}

\item{properties}{No description}

\item{delete_at}{No description}

\item{file_names}{No description}

\item{trash_at}{No description}

\item{is_trashed}{No description}
}
\description{
Collection Object
}
\details{
Todo: Update description
Collection
}

