% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collection_create}
\alias{collection_create}
\title{collection_create}
\description{
Create Arvados collection
}
\section{Fields}{

\describe{
\item{\code{body}}{Structure of the collection we want to create.}
}}

\examples{
arv = Arvados("token", "host_name")
arv$collection_create(list(collection = list(name = "myCollection")))
}
