% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collection_delete}
\alias{collection_delete}
\title{collection_delete}
\description{
Delete Arvados collection

Delete Arvados collection
}
\section{Fields}{

\describe{
\item{\code{uuid}}{UUID of the collection we want to delete.}

\item{\code{uuid}}{UUID of the collection we want to delete.}
}}

\examples{
arv = Arvados("token", "host_name")
arv$collection_delete(uuid = "aaaaa-bbbbb-ccccccccccccccc")
arv = Arvados("token", "host_name")
arv$collection_delete(uuid = "aaaaa-bbbbb-ccccccccccccccc")
}
