% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collection_get}
\alias{collection_get}
\title{collection_get}
\description{
Get Arvados collection
}
\section{Fields}{

\describe{
\item{\code{uuid}}{UUID of the given collection}
}}

\examples{
arv = Arvados("token", "host_name")
arv$collection_get("uuid")
}
