% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collection_list}
\alias{collection_list}
\title{collection_list}
\description{
Retreive list of collections based on provided filter.
}
\section{Fields}{

\describe{
\item{\code{filters}}{List of filters we want to use to retreive list of collections.}

\item{\code{limit}}{Limits the number of result returned by server.}

\item{\code{offset}}{Offset from beginning of the result set.}
}}

\examples{
arv = Arvados("token", "host_name")
arv$collection_list(list("uuid", "=" "aaaaa-bbbbb-ccccccccccccccc"))
}
