% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{CollectionList}
\alias{CollectionList}
\title{CollectionList}
\description{
Collection list
}
\section{Usage}{

\preformatted{collectionList -> CollectionList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#collectionList.}
    \item{etag}{List version.}
    \item{items}{The list of Collections.}
    \item{next_link}{A link to the next page of Collections.}
    \item{next_page_token}{The page token for the next page of Collections.}
    \item{selfLink}{A link back to this list.}
  }
}

