% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.create}
\alias{collections.create}
\title{collections.create}
\usage{
arv$collections.create(collection,
	ensure_unique_name = "false")
}
\arguments{
\item{collection}{Collection object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Collection object.
}
\description{
collections.create is a method defined in Arvados class.
}
