% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.delete}
\alias{collections.delete}
\title{collections.delete}
\usage{
arv$collections.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Collection in question.}
}
\value{
Collection object.
}
\description{
collections.delete is a method defined in Arvados class.
}
