% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.destroy}
\alias{collections.destroy}
\title{collections.destroy is a method defined in Arvados class.}
\usage{
arv$collections.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.destroy is a method defined in Arvados class.
}
