% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.get}
\alias{collections.get}
\title{collections.get}
\usage{
arv$collections.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Collection in question.}
}
\value{
Collection object.
}
\description{
collections.get is a method defined in Arvados class.
}
