% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.provenance}
\alias{collections.provenance}
\title{collections.provenance}
\usage{
arv$collections.provenance(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.provenance is a method defined in Arvados class.
}
