% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.show}
\alias{collections.show}
\title{collections.show is a method defined in Arvados class.}
\usage{
arv$collections.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.show is a method defined in Arvados class.
}
