% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.trash}
\alias{collections.trash}
\title{collections.trash}
\usage{
arv$collections.trash(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.trash is a method defined in Arvados class.
}
