% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.untrash}
\alias{collections.untrash}
\title{collections.untrash}
\usage{
arv$collections.untrash(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.untrash is a method defined in Arvados class.
}
