% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections.used_by}
\alias{collections.used_by}
\title{collections.used_by}
\usage{
arv$collections.used_by(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Collection object.
}
\description{
collections.used_by is a method defined in Arvados class.
}
