% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_create}
\alias{collections_create}
\title{collections_create}
\usage{
arv$collections_create(name, description, owner_uuid = NULL, properties = NULL,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{name}{Name of the collection}

\item{description}{Description of the collection}

\item{owner_uuid}{UUID of the maternal project to created one.}

\item{properties}{Properties of the collection}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
Collection object.
}
\description{
collections_create is a method defined in Arvados class that enables collections creation.
}
