% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_delete}
\alias{collections_delete}
\title{collections_delete}
\usage{
arv$collections_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Collection in question.}
}
\value{
Collection object.
}
\description{
collections_delete is a method defined in Arvados class.
}
