% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_get}
\alias{collections_get}
\title{collections_get}
\usage{
arv$collections_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Collection in question.}
}
\value{
Collection object.
}
\description{
collections_get is a method defined in Arvados class.
}
