% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_list}
\alias{collections_list}
\title{collections_list}
\usage{
arv$collections_list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", cluster_id = NULL, bypass_federation = NULL,
	include_trash = NULL, include_old_versions = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}

\item{include_trash}{Include collections whose is_trashed attribute is true.}

\item{include_old_versions}{Include past collection versions.}
}
\value{
CollectionList object.
}
\description{
collections_list is a method defined in Arvados class.
}
