% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_provenance}
\alias{collections_provenance}
\title{collections_provenance}
\usage{
arv$collections_provenance(uuid)
}
\value{
Collection object.
}
\description{
collections_provenance is a method defined in Arvados class.
}
