% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_trash}
\alias{collections_trash}
\title{collections_trash}
\usage{
arv$collections_trash(uuid)
}
\value{
Collection object.
}
\description{
collections_trash is a method defined in Arvados class.
}
