% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_untrash}
\alias{collections_untrash}
\title{collections_untrash}
\usage{
arv$collections_untrash(uuid)
}
\value{
Collection object.
}
\description{
collections_untrash is a method defined in Arvados class.
}
