% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_update}
\alias{collections_update}
\title{collections_update}
\usage{
arv$collections_update(collection,
	uuid)
}
\arguments{
\item{collection}{Collection object.}

\item{uuid}{The UUID of the Collection in question.}
}
\value{
Collection object.
}
\description{
collections_update is a method defined in Arvados class.
}
