% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{collections_used_by}
\alias{collections_used_by}
\title{collections_used_by}
\usage{
arv$collections_used_by(uuid)
}
\value{
Collection object.
}
\description{
collections_used_by is a method defined in Arvados class.
}
