% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectionTree.R
\docType{data}
\name{CollectionTree}
\alias{CollectionTree}
\title{Arvados Collection Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CollectionTree
}
\description{
Update description
}
\examples{
arv = Collection$new(api, uuid)
}
\keyword{datasets}
