% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosFile.R
\name{connection}
\alias{connection}
\title{connection
Get connection opened in "read" or "write" mode.}
\usage{
connection(rw)
}
\value{
Collection object.
}
\description{
connection
Get connection opened in "read" or "write" mode.
}
