% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Container}
\alias{Container}
\title{Container}
\description{
Container
}
\section{Usage}{

\preformatted{container -> Container$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, state = NULL, started_at = NULL,
	finished_at = NULL, log = NULL, environment = NULL,
	cwd = NULL, command = NULL, output_path = NULL, mounts = NULL,
	runtime_constraints = NULL, output = NULL, container_image = NULL,
	progress = NULL, priority = NULL, updated_at = NULL,
	exit_code = NULL, auth_uuid = NULL, locked_by_uuid = NULL,
	scheduling_parameters = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{state}{}
    \item{started_at}{}
    \item{finished_at}{}
    \item{log}{}
    \item{environment}{}
    \item{cwd}{}
    \item{command}{}
    \item{output_path}{}
    \item{mounts}{}
    \item{runtime_constraints}{}
    \item{output}{}
    \item{container_image}{}
    \item{progress}{}
    \item{priority}{}
    \item{updated_at}{}
    \item{exit_code}{}
    \item{auth_uuid}{}
    \item{locked_by_uuid}{}
    \item{scheduling_parameters}{}
  }
}

