% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests.create}
\alias{container_requests.create}
\title{container_requests.create}
\usage{
arv$container_requests.create(containerrequest,
	ensure_unique_name = "false")
}
\arguments{
\item{containerRequest}{ContainerRequest object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
ContainerRequest object.
}
\description{
container_requests.create is a method defined in Arvados class.
}
