% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests.show}
\alias{container_requests.show}
\title{container_requests.show is a method defined in Arvados class.}
\usage{
arv$container_requests.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
ContainerRequest object.
}
\description{
container_requests.show is a method defined in Arvados class.
}
