% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests.update}
\alias{container_requests.update}
\title{container_requests.update}
\usage{
arv$container_requests.update(containerrequest,
	uuid)
}
\arguments{
\item{containerRequest}{ContainerRequest object.}

\item{uuid}{The UUID of the ContainerRequest in question.}
}
\value{
ContainerRequest object.
}
\description{
container_requests.update is a method defined in Arvados class.
}
