% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests_create}
\alias{container_requests_create}
\title{container_requests_create}
\usage{
arv$container_requests_create(containerrequest,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{containerRequest}{ContainerRequest object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
ContainerRequest object.
}
\description{
container_requests_create is a method defined in Arvados class.
}
