% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests_delete}
\alias{container_requests_delete}
\title{container_requests_delete}
\usage{
arv$container_requests_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the ContainerRequest in question.}
}
\value{
ContainerRequest object.
}
\description{
container_requests_delete is a method defined in Arvados class.
}
