% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests_get}
\alias{container_requests_get}
\title{container_requests_get}
\usage{
arv$container_requests_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the ContainerRequest in question.}
}
\value{
ContainerRequest object.
}
\description{
container_requests_get is a method defined in Arvados class.
}
