% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{container_requests_list}
\alias{container_requests_list}
\title{container_requests_list}
\usage{
arv$container_requests_list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", cluster_id = NULL, bypass_federation = NULL,
	include_trash = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}

\item{include_trash}{Include container requests whose owner project is trashed.}
}
\value{
ContainerRequestList object.
}
\description{
container_requests_list is a method defined in Arvados class.
}
