% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ContainerList}
\alias{ContainerList}
\title{ContainerList}
\description{
Container list
}
\section{Usage}{

\preformatted{containerList -> ContainerList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#containerList.}
    \item{etag}{List version.}
    \item{items}{The list of Containers.}
    \item{next_link}{A link to the next page of Containers.}
    \item{next_page_token}{The page token for the next page of Containers.}
    \item{selfLink}{A link back to this list.}
  }
}

