% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ContainerRequest}
\alias{ContainerRequest}
\title{ContainerRequest}
\description{
ContainerRequest
}
\section{Usage}{

\preformatted{containerRequest -> ContainerRequest$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, name = NULL, description = NULL,
	properties = NULL, state = NULL, requesting_container_uuid = NULL,
	container_uuid = NULL, container_count_max = NULL,
	mounts = NULL, runtime_constraints = NULL, container_image = NULL,
	environment = NULL, cwd = NULL, command = NULL, output_path = NULL,
	priority = NULL, expires_at = NULL, filters = NULL,
	updated_at = NULL, container_count = NULL, use_existing = NULL,
	scheduling_parameters = NULL, output_uuid = NULL, log_uuid = NULL,
	output_name = NULL, output_ttl = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{name}{}
    \item{description}{}
    \item{properties}{}
    \item{state}{}
    \item{requesting_container_uuid}{}
    \item{container_uuid}{}
    \item{container_count_max}{}
    \item{mounts}{}
    \item{runtime_constraints}{}
    \item{container_image}{}
    \item{environment}{}
    \item{cwd}{}
    \item{command}{}
    \item{output_path}{}
    \item{priority}{}
    \item{expires_at}{}
    \item{filters}{}
    \item{updated_at}{}
    \item{container_count}{}
    \item{use_existing}{}
    \item{scheduling_parameters}{}
    \item{output_uuid}{}
    \item{log_uuid}{}
    \item{output_name}{}
    \item{output_ttl}{}
  }
}

