% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{ContainerRequestList}
\alias{ContainerRequestList}
\title{ContainerRequestList}
\description{
ContainerRequest list
}
\section{Usage}{

\preformatted{containerRequestList -> ContainerRequestList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#containerRequestList.}
    \item{etag}{List version.}
    \item{items}{The list of ContainerRequests.}
    \item{next_link}{A link to the next page of ContainerRequests.}
    \item{next_page_token}{The page token for the next page of ContainerRequests.}
    \item{selfLink}{A link back to this list.}
  }
}

