% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.auth}
\alias{containers.auth}
\title{containers.auth}
\usage{
arv$containers.auth(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.auth is a method defined in Arvados class.
}
