% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.create}
\alias{containers.create}
\title{containers.create}
\usage{
arv$containers.create(container,
	ensure_unique_name = "false")
}
\arguments{
\item{container}{Container object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Container object.
}
\description{
containers.create is a method defined in Arvados class.
}
