% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.current}
\alias{containers.current}
\title{containers.current}
\usage{
arv$containers.current(NULL)
}
\value{
Container object.
}
\description{
containers.current is a method defined in Arvados class.
}
