% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.delete}
\alias{containers.delete}
\title{containers.delete}
\usage{
arv$containers.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Container in question.}
}
\value{
Container object.
}
\description{
containers.delete is a method defined in Arvados class.
}
