% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.destroy}
\alias{containers.destroy}
\title{containers.destroy is a method defined in Arvados class.}
\usage{
arv$containers.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.destroy is a method defined in Arvados class.
}
