% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.get}
\alias{containers.get}
\title{containers.get}
\usage{
arv$containers.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Container in question.}
}
\value{
Container object.
}
\description{
containers.get is a method defined in Arvados class.
}
