% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.lock}
\alias{containers.lock}
\title{containers.lock is a method defined in Arvados class.}
\usage{
arv$containers.lock(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.lock is a method defined in Arvados class.
}
