% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.secret_mounts}
\alias{containers.secret_mounts}
\title{containers.secret_mounts}
\usage{
arv$containers.secret_mounts(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.secret_mounts is a method defined in Arvados class.
}
