% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.show}
\alias{containers.show}
\title{containers.show is a method defined in Arvados class.}
\usage{
arv$containers.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.show is a method defined in Arvados class.
}
