% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.unlock}
\alias{containers.unlock}
\title{containers.unlock}
\usage{
arv$containers.unlock(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Container object.
}
\description{
containers.unlock is a method defined in Arvados class.
}
