% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers.update}
\alias{containers.update}
\title{containers.update}
\usage{
arv$containers.update(container,
	uuid)
}
\arguments{
\item{container}{Container object.}

\item{uuid}{The UUID of the Container in question.}
}
\value{
Container object.
}
\description{
containers.update is a method defined in Arvados class.
}
