% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers_current}
\alias{containers_current}
\title{containers_current}
\usage{
arv$containers_current(NULL)
}
\value{
Container object.
}
\description{
containers_current is a method defined in Arvados class.
}
