% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers_delete}
\alias{containers_delete}
\title{containers_delete}
\usage{
arv$containers_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Container in question.}
}
\value{
Container object.
}
\description{
containers_delete is a method defined in Arvados class.
}
