% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers_get}
\alias{containers_get}
\title{containers_get}
\usage{
arv$containers_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Container in question.}
}
\value{
Container object.
}
\description{
containers_get is a method defined in Arvados class.
}
