% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers_secret_mounts}
\alias{containers_secret_mounts}
\title{containers_secret_mounts}
\usage{
arv$containers_secret_mounts(uuid)
}
\value{
Container object.
}
\description{
containers_secret_mounts is a method defined in Arvados class.
}
