% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{containers_unlock}
\alias{containers_unlock}
\title{containers_unlock}
\usage{
arv$containers_unlock(uuid)
}
\value{
Container object.
}
\description{
containers_unlock is a method defined in Arvados class.
}
