% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{copy}
\alias{copy}
\title{copy}
\usage{
copy(content, destination)
}
\arguments{
\item{content}{sth to be copied}

\item{destination}{path to copy sth}
}
\value{
Collection object.
}
\description{
Copies ArvadosFile or Subcollection to another location in the collection.
}
