% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\docType{data}
\name{CTest}
\alias{CTest}
\title{Arvados Collection Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CTest
}
\description{
Update description
}
\examples{
arv = Collection$new(api, uuid)
}
\keyword{datasets}
