% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R, R/Subcollection.R
\name{get}
\alias{get}
\title{get}
\usage{
get(relativePath)

get(relativePath)
}
\arguments{
\item{relativePath}{path from sth is taken}
}
\value{
Collection object.

Collection object.
}
\description{
If relativePath is valid, returns ArvadosFile or Subcollection specified by relativePath, else returns NULL.

If relativePath is valid, returns ArvadosFile or Subcollection specified by relativePath, else returns NULL.
}
