% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subcollection.R
\name{getFileListing()}
\alias{getFileListing()}
\title{getFileListing()}
\usage{
getFileListing()(fullPath)
}
\arguments{
\item{fullPath}{checking if TRUE}
}
\value{
Collection object.
}
\description{
Returns subcollections file content as character vector.
}
