% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosFile.R, R/Subcollection.R
\name{getRelativePath}
\alias{getRelativePath}
\title{getRelativePath}
\usage{
getRelativePath()

getRelativePath()
}
\value{
File path relative to the root

File path relative to the root
}
\description{
Returns file path relative to the root.

Returns subcollection path relative to the root.
}
