% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosFile.R, R/Subcollection.R
\name{getSizeInBytes}
\alias{getSizeInBytes}
\title{getSizeInBytes}
\usage{
getSizeInBytes()

getSizeInBytes()
}
\value{
Collection object.

Collection object.
}
\description{
Returns file size in bytes.

Returns subcollections content size in bytes.
}
