% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Group}
\alias{Group}
\title{Group}
\description{
Group
}
\section{Usage}{

\preformatted{group -> Group$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, created_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, name = NULL,
	description = NULL, updated_at = NULL, group_class = NULL,
	trash_at = NULL, is_trashed = NULL, delete_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{name}{}
    \item{description}{}
    \item{updated_at}{}
    \item{group_class}{}
    \item{trash_at}{}
    \item{is_trashed}{}
    \item{delete_at}{}
  }
}

