% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{GroupList}
\alias{GroupList}
\title{GroupList}
\description{
Group list
}
\section{Usage}{

\preformatted{groupList -> GroupList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#groupList.}
    \item{etag}{List version.}
    \item{items}{The list of Groups.}
    \item{next_link}{A link to the next page of Groups.}
    \item{next_page_token}{The page token for the next page of Groups.}
    \item{selfLink}{A link back to this list.}
  }
}

