% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.contents}
\alias{groups.contents}
\title{groups.contents}
\usage{
arv$groups.contents(filters = NULL,
	where = NULL, order = NULL, distinct = NULL,
	limit = "100", offset = "0", count = "exact",
	include_trash = NULL, uuid = NULL, recursive = NULL)
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}

\item{include_trash}{Include items whose is_trashed attribute is true.}

\item{uuid}{}

\item{recursive}{Include contents from child groups recursively.}
}
\value{
Group object.
}
\description{
groups.contents is a method defined in Arvados class.
}
