% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.create}
\alias{groups.create}
\title{groups.create}
\usage{
arv$groups.create(group, ensure_unique_name = "false")
}
\arguments{
\item{group}{Group object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Group object.
}
\description{
groups.create is a method defined in Arvados class.
}
