% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.delete}
\alias{groups.delete}
\title{groups.delete}
\usage{
arv$groups.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
groups.delete is a method defined in Arvados class.
}
