% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.destroy}
\alias{groups.destroy}
\title{groups.destroy is a method defined in Arvados class.}
\usage{
arv$groups.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Group object.
}
\description{
groups.destroy is a method defined in Arvados class.
}
