% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.get}
\alias{groups.get}
\title{groups.get}
\usage{
arv$groups.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
groups.get is a method defined in Arvados class.
}
