% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.list}
\alias{groups.list}
\title{groups.list}
\usage{
arv$groups.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", include_trash = NULL)
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{select}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}

\item{include_trash}{Include items whose is_trashed attribute is true.}
}
\value{
GroupList object.
}
\description{
groups.list is a method defined in Arvados class.
}
