% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.show}
\alias{groups.show}
\title{groups.show is a method defined in Arvados class.}
\usage{
arv$groups.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Group object.
}
\description{
groups.show is a method defined in Arvados class.
}
