% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.trash}
\alias{groups.trash}
\title{groups.trash}
\usage{
arv$groups.trash(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Group object.
}
\description{
groups.trash is a method defined in Arvados class.
}
