% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.untrash}
\alias{groups.untrash}
\title{groups.untrash}
\usage{
arv$groups.untrash(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Group object.
}
\description{
groups.untrash is a method defined in Arvados class.
}
