% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups.update}
\alias{groups.update}
\title{groups.update}
\usage{
arv$groups.update(group, uuid)
}
\arguments{
\item{group}{Group object.}

\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
groups.update is a method defined in Arvados class.
}
