% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_contents}
\alias{groups_contents}
\title{groups_contents}
\usage{
arv$groups_contents(filters = NULL,
	where = NULL, order = NULL, distinct = NULL,
	limit = "100", offset = "0", count = "exact",
	cluster_id = NULL, bypass_federation = NULL,
	include_trash = NULL, uuid = NULL, recursive = NULL,
	include = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}

\item{include_trash}{Include items whose is_trashed attribute is true.}

\item{recursive}{Include contents from child groups recursively.}

\item{include}{Include objects referred to by listed field in "included" (only owner_uuid)}
}
\value{
Group object.
}
\description{
groups_contents is a method defined in Arvados class.
}
