% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_create}
\alias{groups_create}
\title{groups_create}
\usage{
arv$groups_create(group, ensure_unique_name = "false",
	cluster_id = NULL, async = "false")
}
\arguments{
\item{group}{Group object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}

\item{async}{Defer permissions update.}
}
\value{
Group object.
}
\description{
groups_create is a method defined in Arvados class that supports project creation.
}
