% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_delete}
\alias{groups_delete}
\title{groups_delete}
\usage{
arv$groups_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
groups_delete is a method defined in Arvados class.
}
