% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_get}
\alias{groups_get}
\title{groups_get}
\usage{
arv$groups_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
groups_get is a method defined in Arvados class.
}
