% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_shared}
\alias{groups_shared}
\title{groups_shared}
\usage{
arv$groups_shared(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", cluster_id = NULL, bypass_federation = NULL,
	include_trash = NULL, include = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}

\item{include_trash}{Include items whose is_trashed attribute is true.}
}
\value{
Group object.
}
\description{
groups_shared is a method defined in Arvados class.
}
