% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{groups_update}
\alias{groups_update}
\title{groups_update}
\usage{
arv$groups_update(group, uuid,
	async = "false")
}
\arguments{
\item{group}{Group object.}

\item{uuid}{The UUID of the Group in question.}

\item{async}{defer permissions update}
}
\value{
Group object.
}
\description{
groups_update is a method defined in Arvados class.
}
