% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Human}
\alias{Human}
\title{Human}
\description{
Human
}
\section{Usage}{

\preformatted{human -> Human$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, properties = NULL,
	created_at = NULL, updated_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{properties}{}
    \item{created_at}{}
    \item{updated_at}{}
  }
}

