% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{HumanList}
\alias{HumanList}
\title{HumanList}
\description{
Human list
}
\section{Usage}{

\preformatted{humanList -> HumanList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#humanList.}
    \item{etag}{List version.}
    \item{items}{The list of Humans.}
    \item{next_link}{A link to the next page of Humans.}
    \item{next_page_token}{The page token for the next page of Humans.}
    \item{selfLink}{A link back to this list.}
  }
}

