% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.create}
\alias{humans.create}
\title{humans.create}
\usage{
arv$humans.create(human, ensure_unique_name = "false")
}
\arguments{
\item{human}{Human object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Human object.
}
\description{
humans.create is a method defined in Arvados class.
}
