% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.delete}
\alias{humans.delete}
\title{humans.delete}
\usage{
arv$humans.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Human in question.}
}
\value{
Human object.
}
\description{
humans.delete is a method defined in Arvados class.
}
