% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.destroy}
\alias{humans.destroy}
\title{humans.destroy is a method defined in Arvados class.}
\usage{
arv$humans.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Human object.
}
\description{
humans.destroy is a method defined in Arvados class.
}
