% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.get}
\alias{humans.get}
\title{humans.get}
\usage{
arv$humans.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Human in question.}
}
\value{
Human object.
}
\description{
humans.get is a method defined in Arvados class.
}
