% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.list}
\alias{humans.list}
\title{humans.list}
\usage{
arv$humans.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\value{
HumanList object.
}
\description{
humans.list is a method defined in Arvados class.
}
