% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.show}
\alias{humans.show}
\title{humans.show is a method defined in Arvados class.}
\usage{
arv$humans.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Human object.
}
\description{
humans.show is a method defined in Arvados class.
}
