% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{humans.update}
\alias{humans.update}
\title{humans.update}
\usage{
arv$humans.update(human, uuid)
}
\arguments{
\item{human}{Human object.}

\item{uuid}{The UUID of the Human in question.}
}
\value{
Human object.
}
\description{
humans.update is a method defined in Arvados class.
}
