% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Job}
\alias{Job}
\title{Job}
\description{
Job
}
\section{Usage}{

\preformatted{job -> Job$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, submit_id = NULL,
	script = NULL, script_version = NULL, script_parameters = NULL,
	cancelled_by_client_uuid = NULL, cancelled_by_user_uuid = NULL,
	cancelled_at = NULL, started_at = NULL, finished_at = NULL,
	running = NULL, success = NULL, output = NULL, created_at = NULL,
	updated_at = NULL, is_locked_by_uuid = NULL, log = NULL,
	tasks_summary = NULL, runtime_constraints = NULL, nondeterministic = NULL,
	repository = NULL, supplied_script_version = NULL,
	docker_image_locator = NULL, priority = NULL, description = NULL,
	state = NULL, arvados_sdk_version = NULL, components = NULL,
	script_parameters_digest = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{submit_id}{}
    \item{script}{}
    \item{script_version}{}
    \item{script_parameters}{}
    \item{cancelled_by_client_uuid}{}
    \item{cancelled_by_user_uuid}{}
    \item{cancelled_at}{}
    \item{started_at}{}
    \item{finished_at}{}
    \item{running}{}
    \item{success}{}
    \item{output}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{is_locked_by_uuid}{}
    \item{log}{}
    \item{tasks_summary}{}
    \item{runtime_constraints}{}
    \item{nondeterministic}{}
    \item{repository}{}
    \item{supplied_script_version}{}
    \item{docker_image_locator}{}
    \item{priority}{}
    \item{description}{}
    \item{state}{}
    \item{arvados_sdk_version}{}
    \item{components}{}
    \item{script_parameters_digest}{}
  }
}

