% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{job_tasks.create}
\alias{job_tasks.create}
\title{job_tasks.create}
\usage{
arv$job_tasks.create(jobtask, ensure_unique_name = "false")
}
\arguments{
\item{jobTask}{JobTask object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
JobTask object.
}
\description{
job_tasks.create is a method defined in Arvados class.
}
