% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{job_tasks.delete}
\alias{job_tasks.delete}
\title{job_tasks.delete}
\usage{
arv$job_tasks.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the JobTask in question.}
}
\value{
JobTask object.
}
\description{
job_tasks.delete is a method defined in Arvados class.
}
