% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{job_tasks.show}
\alias{job_tasks.show}
\title{job_tasks.show is a method defined in Arvados class.}
\usage{
arv$job_tasks.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
JobTask object.
}
\description{
job_tasks.show is a method defined in Arvados class.
}
