% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{job_tasks.update}
\alias{job_tasks.update}
\title{job_tasks.update}
\usage{
arv$job_tasks.update(jobtask, uuid)
}
\arguments{
\item{jobTask}{JobTask object.}

\item{uuid}{The UUID of the JobTask in question.}
}
\value{
JobTask object.
}
\description{
job_tasks.update is a method defined in Arvados class.
}
