% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{JobList}
\alias{JobList}
\title{JobList}
\description{
Job list
}
\section{Usage}{

\preformatted{jobList -> JobList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#jobList.}
    \item{etag}{List version.}
    \item{items}{The list of Jobs.}
    \item{next_link}{A link to the next page of Jobs.}
    \item{next_page_token}{The page token for the next page of Jobs.}
    \item{selfLink}{A link back to this list.}
  }
}

