% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.create}
\alias{jobs.create}
\title{jobs.create}
\usage{
arv$jobs.create(job, ensure_unique_name = "false",
	find_or_create = "false", filters = NULL,
	minimum_script_version = NULL, exclude_script_versions = NULL)
}
\arguments{
\item{job}{Job object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}

\item{find_or_create}{}

\item{filters}{}

\item{minimum_script_version}{}

\item{exclude_script_versions}{}
}
\value{
Job object.
}
\description{
jobs.create is a method defined in Arvados class.
}
