% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.delete}
\alias{jobs.delete}
\title{jobs.delete}
\usage{
arv$jobs.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Job in question.}
}
\value{
Job object.
}
\description{
jobs.delete is a method defined in Arvados class.
}
