% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.destroy}
\alias{jobs.destroy}
\title{jobs.destroy is a method defined in Arvados class.}
\usage{
arv$jobs.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Job object.
}
\description{
jobs.destroy is a method defined in Arvados class.
}
