% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.get}
\alias{jobs.get}
\title{jobs.get}
\usage{
arv$jobs.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Job in question.}
}
\value{
Job object.
}
\description{
jobs.get is a method defined in Arvados class.
}
