% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.list}
\alias{jobs.list}
\title{jobs.list}
\usage{
arv$jobs.list(filters = NULL, where = NULL,
	order = NULL, select = NULL, distinct = NULL,
	limit = "100", offset = "0", count = "exact")
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{select}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}
}
\value{
JobList object.
}
\description{
jobs.list is a method defined in Arvados class.
}
