% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.show}
\alias{jobs.show}
\title{jobs.show is a method defined in Arvados class.}
\usage{
arv$jobs.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Job object.
}
\description{
jobs.show is a method defined in Arvados class.
}
