% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{jobs.update}
\alias{jobs.update}
\title{jobs.update}
\usage{
arv$jobs.update(job, uuid)
}
\arguments{
\item{job}{Job object.}

\item{uuid}{The UUID of the Job in question.}
}
\value{
Job object.
}
\description{
jobs.update is a method defined in Arvados class.
}
