% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{JobTask}
\alias{JobTask}
\title{JobTask}
\description{
JobTask
}
\section{Usage}{

\preformatted{jobTask -> JobTask$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, job_uuid = NULL,
	sequence = NULL, parameters = NULL, output = NULL,
	progress = NULL, success = NULL, created_at = NULL,
	updated_at = NULL, created_by_job_task_uuid = NULL,
	qsequence = NULL, started_at = NULL, finished_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{job_uuid}{}
    \item{sequence}{}
    \item{parameters}{}
    \item{output}{}
    \item{progress}{}
    \item{success}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{created_by_job_task_uuid}{}
    \item{qsequence}{}
    \item{started_at}{}
    \item{finished_at}{}
  }
}

