% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{JobTaskList}
\alias{JobTaskList}
\title{JobTaskList}
\description{
JobTask list
}
\section{Usage}{

\preformatted{jobTaskList -> JobTaskList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#jobTaskList.}
    \item{etag}{List version.}
    \item{items}{The list of JobTasks.}
    \item{next_link}{A link to the next page of JobTasks.}
    \item{next_page_token}{The page token for the next page of JobTasks.}
    \item{selfLink}{A link back to this list.}
  }
}

