% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_disks.create}
\alias{keep_disks.create}
\title{keep_disks.create}
\usage{
arv$keep_disks.create(keepdisk,
	ensure_unique_name = "false")
}
\arguments{
\item{keepDisk}{KeepDisk object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
KeepDisk object.
}
\description{
keep_disks.create is a method defined in Arvados class.
}
