% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_disks.destroy}
\alias{keep_disks.destroy}
\title{keep_disks.destroy is a method defined in Arvados class.}
\usage{
arv$keep_disks.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
KeepDisk object.
}
\description{
keep_disks.destroy is a method defined in Arvados class.
}
