% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_disks.get}
\alias{keep_disks.get}
\title{keep_disks.get}
\usage{
arv$keep_disks.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the KeepDisk in question.}
}
\value{
KeepDisk object.
}
\description{
keep_disks.get is a method defined in Arvados class.
}
