% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_disks.update}
\alias{keep_disks.update}
\title{keep_disks.update is a method defined in Arvados class.}
\usage{
arv$keep_disks.update(keep_disk,
	uuid)
}
\arguments{
\item{keepDisk}{KeepDisk object.}

\item{uuid}{The UUID of the KeepDisk in question.}
}
\value{
KeepDisk object.
}
\description{
keep_disks.update is a method defined in Arvados class.
}
