% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.create}
\alias{keep_services.create}
\title{keep_services.create}
\usage{
arv$keep_services.create(keepservice,
	ensure_unique_name = "false")
}
\arguments{
\item{keepService}{KeepService object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
KeepService object.
}
\description{
keep_services.create is a method defined in Arvados class.
}
