% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.destroy}
\alias{keep_services.destroy}
\title{keep_services.destroy is a method defined in Arvados class.}
\usage{
arv$keep_services.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
KeepService object.
}
\description{
keep_services.destroy is a method defined in Arvados class.
}
