% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.get}
\alias{keep_services.get}
\title{keep_services.get}
\usage{
arv$keep_services.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the KeepService in question.}
}
\value{
KeepService object.
}
\description{
keep_services.get is a method defined in Arvados class.
}
