% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.list}
\alias{keep_services.list}
\title{keep_services.list}
\usage{
arv$keep_services.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{select}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}
}
\value{
KeepServiceList object.
}
\description{
keep_services.list is a method defined in Arvados class.
}
