% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.show}
\alias{keep_services.show}
\title{keep_services.show is a method defined in Arvados class.}
\usage{
arv$keep_services.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
KeepService object.
}
\description{
keep_services.show is a method defined in Arvados class.
}
