% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services.update}
\alias{keep_services.update}
\title{keep_services.update}
\usage{
arv$keep_services.update(keepservice,
	uuid)
}
\arguments{
\item{keepService}{KeepService object.}

\item{uuid}{The UUID of the KeepService in question.}
}
\value{
KeepService object.
}
\description{
keep_services.update is a method defined in Arvados class.
}
