% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services_accessible}
\alias{keep_services_accessible}
\title{keep_services_accessible}
\usage{
arv$keep_services_accessible(NULL)
}
\value{
KeepService object.
}
\description{
keep_services_accessible is a method defined in Arvados class.
}
