% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{keep_services_create}
\alias{keep_services_create}
\title{keep_services_create}
\usage{
arv$keep_services_create(keepservice,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{keepService}{KeepService object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
KeepService object.
}
\description{
keep_services_create is a method defined in Arvados class.
}
