% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{KeepDisk}
\alias{KeepDisk}
\title{KeepDisk}
\description{
KeepDisk
}
\section{Usage}{

\preformatted{keepDisk -> KeepDisk$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, ping_secret = NULL,
	node_uuid = NULL, filesystem_uuid = NULL, bytes_total = NULL,
	bytes_free = NULL, is_readable = NULL, is_writable = NULL,
	last_read_at = NULL, last_write_at = NULL, last_ping_at = NULL,
	created_at = NULL, updated_at = NULL, keep_service_uuid = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{ping_secret}{}
    \item{node_uuid}{}
    \item{filesystem_uuid}{}
    \item{bytes_total}{}
    \item{bytes_free}{}
    \item{is_readable}{}
    \item{is_writable}{}
    \item{last_read_at}{}
    \item{last_write_at}{}
    \item{last_ping_at}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{keep_service_uuid}{}
  }
}

