% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{KeepDiskList}
\alias{KeepDiskList}
\title{KeepDiskList}
\description{
KeepDisk list
}
\section{Usage}{

\preformatted{keepDiskList -> KeepDiskList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#keepDiskList.}
    \item{etag}{List version.}
    \item{items}{The list of KeepDisks.}
    \item{next_link}{A link to the next page of KeepDisks.}
    \item{next_page_token}{The page token for the next page of KeepDisks.}
    \item{selfLink}{A link back to this list.}
  }
}

