% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{KeepService}
\alias{KeepService}
\title{KeepService}
\description{
KeepService
}
\section{Usage}{

\preformatted{keepService -> KeepService$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, service_host = NULL,
	service_port = NULL, service_ssl_flag = NULL, service_type = NULL,
	created_at = NULL, updated_at = NULL, read_only = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{service_host}{}
    \item{service_port}{}
    \item{service_ssl_flag}{}
    \item{service_type}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{read_only}{}
  }
}

