% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{KeepServiceList}
\alias{KeepServiceList}
\title{KeepServiceList}
\description{
KeepService list
}
\section{Usage}{

\preformatted{keepServiceList -> KeepServiceList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#keepServiceList.}
    \item{etag}{List version.}
    \item{items}{The list of KeepServices.}
    \item{next_link}{A link to the next page of KeepServices.}
    \item{next_page_token}{The page token for the next page of KeepServices.}
    \item{selfLink}{A link back to this list.}
  }
}

