% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Link}
\alias{Link}
\title{Link}
\description{
Link
}
\section{Usage}{

\preformatted{link -> Link$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, created_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, tail_uuid = NULL,
	link_class = NULL, name = NULL, head_uuid = NULL, properties = NULL,
	updated_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{tail_uuid}{}
    \item{link_class}{}
    \item{name}{}
    \item{head_uuid}{}
    \item{properties}{}
    \item{updated_at}{}
  }
}

