% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{LinkList}
\alias{LinkList}
\title{LinkList}
\description{
Link list
}
\section{Usage}{

\preformatted{linkList -> LinkList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#linkList.}
    \item{etag}{List version.}
    \item{items}{The list of Links.}
    \item{next_link}{A link to the next page of Links.}
    \item{next_page_token}{The page token for the next page of Links.}
    \item{selfLink}{A link back to this list.}
  }
}

