% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.create}
\alias{links.create}
\title{links.create}
\usage{
arv$links.create(link, ensure_unique_name = "false")
}
\arguments{
\item{link}{Link object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Link object.
}
\description{
links.create is a method defined in Arvados class.
}
