% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.delete}
\alias{links.delete}
\title{links.delete}
\usage{
arv$links.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links.delete is a method defined in Arvados class.
}
