% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.destroy}
\alias{links.destroy}
\title{links.destroy is a method defined in Arvados class.}
\usage{
arv$links.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Link object.
}
\description{
links.destroy is a method defined in Arvados class.
}
