% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.get}
\alias{links.get}
\title{links.get}
\usage{
arv$links.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links.get is a method defined in Arvados class.
}
