% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.get_permissions}
\alias{links.get_permissions}
\title{links.get_permissions}
\usage{
arv$links.get_permissions(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Link object.
}
\description{
links.get_permissions is a method defined in Arvados class.
}
