% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.show}
\alias{links.show}
\title{links.show is a method defined in Arvados class.}
\usage{
arv$links.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Link object.
}
\description{
links.show is a method defined in Arvados class.
}
