% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links.update}
\alias{links.update}
\title{links.update}
\usage{
arv$links.update(link, uuid)
}
\arguments{
\item{link}{Link object.}

\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links.update is a method defined in Arvados class.
}
