% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links_create}
\alias{links_create}
\title{links_create}
\usage{
arv$links_create(link, ensure_unique_name = "false",
	cluster_id = NULL)
}
\arguments{
\item{link}{Link object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
Link object.
}
\description{
links_create is a method defined in Arvados class.
}
