% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links_delete}
\alias{links_delete}
\title{links_delete}
\usage{
arv$links_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links_delete is a method defined in Arvados class.
}
