% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links_get}
\alias{links_get}
\title{links_get}
\usage{
arv$links_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links_get is a method defined in Arvados class.
}
