% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links_get_permissions}
\alias{links_get_permissions}
\title{links_get_permissions}
\usage{
arv$links_get_permissions(uuid)
}
\value{
Link object.
}
\description{
links_get_permissions is a method defined in Arvados class.
}
