% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{links_update}
\alias{links_update}
\title{links_update}
\usage{
arv$links_update(link, uuid)
}
\arguments{
\item{link}{Link object.}

\item{uuid}{The UUID of the Link in question.}
}
\value{
Link object.
}
\description{
links_update is a method defined in Arvados class.
}
