% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listAll}
\alias{listAll}
\title{listAll}
\usage{
listAll(fn, ...)
}
\arguments{
\item{fn}{Arvados method used to retrieve items from REST service.}

\item{...}{Optional arguments which will be pased to fn .}
}
\description{
List all resources even if the number of items is greater than maximum API limit.
}
\examples{
\dontrun{
arv <- Arvados$new("your Arvados token", "example.arvadosapi.com")
cl <- listAll(arv$collections.list, filters = list(list("name", "like", "test\%"))
}
}
