% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Log}
\alias{Log}
\title{Log}
\description{
Log
}
\section{Usage}{

\preformatted{log -> Log$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, object_uuid = NULL, event_at = NULL,
	event_type = NULL, summary = NULL, properties = NULL,
	created_at = NULL, updated_at = NULL, modified_at = NULL,
	object_owner_uuid = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{object_uuid}{}
    \item{event_at}{}
    \item{event_type}{}
    \item{summary}{}
    \item{properties}{}
    \item{created_at}{}
    \item{updated_at}{}
    \item{modified_at}{}
    \item{object_owner_uuid}{}
  }
}

