% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{LogList}
\alias{LogList}
\title{LogList}
\description{
Log list
}
\section{Usage}{

\preformatted{logList -> LogList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#logList.}
    \item{etag}{List version.}
    \item{items}{The list of Logs.}
    \item{next_link}{A link to the next page of Logs.}
    \item{next_page_token}{The page token for the next page of Logs.}
    \item{selfLink}{A link back to this list.}
  }
}

