% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.create}
\alias{logs.create}
\title{logs.create}
\usage{
arv$logs.create(log, ensure_unique_name = "false")
}
\arguments{
\item{log}{Log object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Log object.
}
\description{
logs.create is a method defined in Arvados class.
}
