% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.delete}
\alias{logs.delete}
\title{logs.delete}
\usage{
arv$logs.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs.delete is a method defined in Arvados class.
}
