% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.destroy}
\alias{logs.destroy}
\title{logs.destroy is a method defined in Arvados class.}
\usage{
arv$logs.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Log object.
}
\description{
logs.destroy is a method defined in Arvados class.
}
