% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.get}
\alias{logs.get}
\title{logs.get}
\usage{
arv$logs.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs.get is a method defined in Arvados class.
}
