% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.index}
\alias{logs.index}
\title{logs.index is a method defined in Arvados class.}
\usage{
arv$logs.index(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{select}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}
}
\value{
LogList object.
}
\description{
logs.index is a method defined in Arvados class.
}
