% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.show}
\alias{logs.show}
\title{logs.show is a method defined in Arvados class.}
\usage{
arv$logs.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Log object.
}
\description{
logs.show is a method defined in Arvados class.
}
