% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs.update}
\alias{logs.update}
\title{logs.update}
\usage{
arv$logs.update(log, uuid)
}
\arguments{
\item{log}{Log object.}

\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs.update is a method defined in Arvados class.
}
