% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs_delete}
\alias{logs_delete}
\title{logs_delete}
\usage{
arv$logs_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs_delete is a method defined in Arvados class.
}
