% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs_get}
\alias{logs_get}
\title{logs_get}
\usage{
arv$logs_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs_get is a method defined in Arvados class.
}
