% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{logs_update}
\alias{logs_update}
\title{logs_update}
\usage{
arv$logs_update(log, uuid)
}
\arguments{
\item{log}{Log object.}

\item{uuid}{The UUID of the Log in question.}
}
\value{
Log object.
}
\description{
logs_update is a method defined in Arvados class.
}
