% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{move}
\alias{move}
\title{move}
\usage{
move(content, destination)
}
\arguments{
\item{content}{sth to be moved}

\item{destination}{path to move sth}
}
\value{
Collection object.
}
\description{
Moves ArvadosFile or Subcollection to another location in the collection.
}
