% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosFile.R, R/Subcollection.R
\name{name}
\alias{name}
\title{remove}
\usage{
remove(name)

move(destination)

copy(destination)

remove(name)

move(destination)

copy(destination)
}
\arguments{
\item{name}{Name of the file to be removed}
}
\value{
Collection object.

Collection object.

Collection object.

Collection object.

Collection object.

Collection object.
}
\description{
Removes ArvadosFile or Subcollection specified by name from the subcollection.

Moves file to a new location inside collection.

Copies file to a new location inside collection.

Removes ArvadosFile or Subcollection specified by name from the subcollection.

Moves subcollection to a new location inside collection.

Copies subcollection to a new location inside collection.
}
