% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Node}
\alias{Node}
\title{Node}
\description{
Node
}
\section{Usage}{

\preformatted{node -> Node$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, created_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, slot_number = NULL,
	hostname = NULL, domain = NULL, ip_address = NULL,
	first_ping_at = NULL, last_ping_at = NULL, info = NULL,
	updated_at = NULL, properties = NULL, job_uuid = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{slot_number}{}
    \item{hostname}{}
    \item{domain}{}
    \item{ip_address}{}
    \item{first_ping_at}{}
    \item{last_ping_at}{}
    \item{info}{}
    \item{updated_at}{}
    \item{properties}{}
    \item{job_uuid}{}
  }
}

