% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{NodeList}
\alias{NodeList}
\title{NodeList}
\description{
Node list
}
\section{Usage}{

\preformatted{nodeList -> NodeList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#nodeList.}
    \item{etag}{List version.}
    \item{items}{The list of Nodes.}
    \item{next_link}{A link to the next page of Nodes.}
    \item{next_page_token}{The page token for the next page of Nodes.}
    \item{selfLink}{A link back to this list.}
  }
}

