% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.create}
\alias{nodes.create}
\title{nodes.create}
\usage{
arv$nodes.create(node, ensure_unique_name = "false",
	assign_slot = NULL)
}
\arguments{
\item{node}{Node object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}

\item{assign_slot}{assign slot and hostname}
}
\value{
Node object.
}
\description{
nodes.create is a method defined in Arvados class.
}
