% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.delete}
\alias{nodes.delete}
\title{nodes.delete}
\usage{
arv$nodes.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Node in question.}
}
\value{
Node object.
}
\description{
nodes.delete is a method defined in Arvados class.
}
