% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.destroy}
\alias{nodes.destroy}
\title{nodes.destroy is a method defined in Arvados class.}
\usage{
arv$nodes.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Node object.
}
\description{
nodes.destroy is a method defined in Arvados class.
}
