% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.get}
\alias{nodes.get}
\title{nodes.get}
\usage{
arv$nodes.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Node in question.}
}
\value{
Node object.
}
\description{
nodes.get is a method defined in Arvados class.
}
