% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.list}
\alias{nodes.list}
\title{nodes.list}
\usage{
arv$nodes.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\value{
NodeList object.
}
\description{
nodes.list is a method defined in Arvados class.
}
