% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.ping}
\alias{nodes.ping}
\title{nodes.ping}
\usage{
arv$nodes.ping(uuid, ping_secret)
}
\arguments{
\item{uuid}{}

\item{ping_secret}{}
}
\value{
Node object.
}
\description{
nodes.ping is a method defined in Arvados class.
}
