% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.show}
\alias{nodes.show}
\title{nodes.show is a method defined in Arvados class.}
\usage{
arv$nodes.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Node object.
}
\description{
nodes.show is a method defined in Arvados class.
}
