% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{nodes.update}
\alias{nodes.update}
\title{nodes.update}
\usage{
arv$nodes.update(node, uuid, assign_slot = NULL)
}
\arguments{
\item{node}{Node object.}

\item{uuid}{The UUID of the Node in question.}

\item{assign_slot}{assign slot and hostname}
}
\value{
Node object.
}
\description{
nodes.update is a method defined in Arvados class.
}
