% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_instances.cancel}
\alias{pipeline_instances.cancel}
\title{pipeline_instances.cancel}
\usage{
arv$pipeline_instances.cancel(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
PipelineInstance object.
}
\description{
pipeline_instances.cancel is a method defined in Arvados class.
}
