% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_instances.get}
\alias{pipeline_instances.get}
\title{pipeline_instances.get}
\usage{
arv$pipeline_instances.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the PipelineInstance in question.}
}
\value{
PipelineInstance object.
}
\description{
pipeline_instances.get is a method defined in Arvados class.
}
