% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_instances.show}
\alias{pipeline_instances.show}
\title{pipeline_instances.show is a method defined in Arvados class.}
\usage{
arv$pipeline_instances.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
PipelineInstance object.
}
\description{
pipeline_instances.show is a method defined in Arvados class.
}
