% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_instances.update}
\alias{pipeline_instances.update}
\title{pipeline_instances.update}
\usage{
arv$pipeline_instances.update(pipelineinstance,
	uuid)
}
\arguments{
\item{pipelineInstance}{PipelineInstance object.}

\item{uuid}{The UUID of the PipelineInstance in question.}
}
\value{
PipelineInstance object.
}
\description{
pipeline_instances.update is a method defined in Arvados class.
}
