% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_templates.create}
\alias{pipeline_templates.create}
\title{pipeline_templates.create}
\usage{
arv$pipeline_templates.create(pipelinetemplate,
	ensure_unique_name = "false")
}
\arguments{
\item{pipelineTemplate}{PipelineTemplate object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
PipelineTemplate object.
}
\description{
pipeline_templates.create is a method defined in Arvados class.
}
