% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_templates.get}
\alias{pipeline_templates.get}
\title{pipeline_templates.get}
\usage{
arv$pipeline_templates.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the PipelineTemplate in question.}
}
\value{
PipelineTemplate object.
}
\description{
pipeline_templates.get is a method defined in Arvados class.
}
