% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_templates.show}
\alias{pipeline_templates.show}
\title{pipeline_templates.show is a method defined in Arvados class.}
\usage{
arv$pipeline_templates.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
PipelineTemplate object.
}
\description{
pipeline_templates.show is a method defined in Arvados class.
}
