% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{pipeline_templates.update}
\alias{pipeline_templates.update}
\title{pipeline_templates.update}
\usage{
arv$pipeline_templates.update(pipelinetemplate,
	uuid)
}
\arguments{
\item{pipelineTemplate}{PipelineTemplate object.}

\item{uuid}{The UUID of the PipelineTemplate in question.}
}
\value{
PipelineTemplate object.
}
\description{
pipeline_templates.update is a method defined in Arvados class.
}
