% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{PipelineInstance}
\alias{PipelineInstance}
\title{PipelineInstance}
\description{
PipelineInstance
}
\section{Usage}{

\preformatted{pipelineInstance -> PipelineInstance$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_by_client_uuid = NULL, modified_by_user_uuid = NULL,
	modified_at = NULL, pipeline_template_uuid = NULL,
	name = NULL, components = NULL, updated_at = NULL,
	properties = NULL, state = NULL, components_summary = NULL,
	started_at = NULL, finished_at = NULL, description = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{pipeline_template_uuid}{}
    \item{name}{}
    \item{components}{}
    \item{updated_at}{}
    \item{properties}{}
    \item{state}{}
    \item{components_summary}{}
    \item{started_at}{}
    \item{finished_at}{}
    \item{description}{}
  }
}

