% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{PipelineInstanceList}
\alias{PipelineInstanceList}
\title{PipelineInstanceList}
\description{
PipelineInstance list
}
\section{Usage}{

\preformatted{pipelineInstanceList -> PipelineInstanceList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#pipelineInstanceList.}
    \item{etag}{List version.}
    \item{items}{The list of PipelineInstances.}
    \item{next_link}{A link to the next page of PipelineInstances.}
    \item{next_page_token}{The page token for the next page of PipelineInstances.}
    \item{selfLink}{A link back to this list.}
  }
}

