% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{PipelineTemplate}
\alias{PipelineTemplate}
\title{PipelineTemplate}
\description{
PipelineTemplate
}
\section{Usage}{

\preformatted{pipelineTemplate -> PipelineTemplate$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_by_client_uuid = NULL, modified_by_user_uuid = NULL,
	modified_at = NULL, name = NULL, components = NULL,
	updated_at = NULL, description = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{name}{}
    \item{components}{}
    \item{updated_at}{}
    \item{description}{}
  }
}

