% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{PipelineTemplateList}
\alias{PipelineTemplateList}
\title{PipelineTemplateList}
\description{
PipelineTemplate list
}
\section{Usage}{

\preformatted{pipelineTemplateList -> PipelineTemplateList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#pipelineTemplateList.}
    \item{etag}{List version.}
    \item{items}{The list of PipelineTemplates.}
    \item{next_link}{A link to the next page of PipelineTemplates.}
    \item{next_page_token}{The page token for the next page of PipelineTemplates.}
    \item{selfLink}{A link back to this list.}
  }
}

