% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{print.Collection}
\alias{print.Collection}
\title{print.Collection}
\usage{
\method{print}{Collection}(x, ...)
}
\arguments{
\item{x}{Instance of Collection class}

\item{...}{Optional arguments.}
}
\description{
Custom print function for Collection class
}
