% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subcollection.R
\name{print.Subcollection}
\alias{print.Subcollection}
\title{print.Subcollection}
\usage{
\method{print}{Subcollection}(x, ...)
}
\arguments{
\item{x}{Instance of Subcollection class}

\item{...}{Optional arguments.}
}
\description{
Custom print function for Subcollection class
}
