% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects.create}
\alias{projects.create}
\title{project.create}
\usage{
arv$projects.create(group, ensure_unique_name = "false")
}
\arguments{
\item{group}{Group object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Group object.
}
\description{
projects.create wrapps groups.create method by setting group_class attribute to "project".
}
