% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects.delete}
\alias{projects.delete}
\title{project.delete}
\usage{
arv$project.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
projects.delete is equivalent to groups.delete method.
}
