% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects.get}
\alias{projects.get}
\title{project.get}
\usage{
arv$projects.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
projects.get is equivalent to groups.get method.
}
