% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects.list}
\alias{projects.list}
\title{project.list}
\usage{
arv$projects.list(filters = NULL,
	where = NULL, order = NULL, distinct = NULL,
	limit = "100", offset = "0", count = "exact",
	include_trash = NULL, uuid = NULL, recursive = NULL)
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}

\item{include_trash}{Include items whose is_trashed attribute is true.}

\item{uuid}{}

\item{recursive}{Include contents from child groups recursively.}
}
\value{
Group object.
}
\description{
projects.list wrapps groups.list method by setting group_class attribute to "project".
}
