% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_create}
\alias{projects_create}
\title{project.create}
\usage{
arv$projects_create(name, description, owner_uuid, properties = NULL, ensure_unique_name = "false")
}
\arguments{
\item{name}{Name of the project.}

\item{description}{Description of the project.}

\item{owner_uuid}{UUID of the maternal project to created one.}

\item{properties}{List of the properties of the project.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}
}
\value{
Group object.
}
\description{
projects_create wrapps groups_create method by setting group_class attribute to "project".
}
