% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_exist}
\alias{projects_exist}
\title{projects_exist}
\usage{
arv$projects_exist(uuid)
}
\arguments{
\item{uuid}{A uuid of a project / file.}
}
\value{
Project object.
}
\description{
projects_exist is a method defined in Arvados class that enables checking id the project with such a UUID exist.
}
