% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_get}
\alias{projects_get}
\title{project.get}
\usage{
arv$projects_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
projects_get is equivalent to groups_get method.
}
