% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_permission_check}
\alias{projects_permission_check}
\title{projects_permission_check}
\usage{
arv$projects_permission_check(object, user, type = NULL)
}
\arguments{
\item{uuid_file}{A uuid of a project / file.}

\item{uuid_person}{A uuid of the person that the permission is being updated.}

\item{type}{Possible options are can_read / can_write / can_manage.}
}
\value{
Link object.
}
\description{
projects_permission_check is a method defined in Arvados class that enables checking file permissions.
}
