% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_permission_give}
\alias{projects_permission_give}
\title{projects_permission_give}
\usage{
arv$projects_permission_give(type, object, user)
}
\arguments{
\item{type}{Possible options are can_read / can_write / can_manage.}

\item{object}{A uuid of a project / file.}

\item{user}{A uuid of the person that gets the permission_}
}
\value{
Granted permissions.
}
\description{
projects_permission_give is a method defined in Arvados class that enables sharing files with another users_
}
