% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_permission_refuse}
\alias{projects_permission_refuse}
\title{projects_permission_refuse}
\usage{
arv$projects_permission_refuse(type, object, user)
}
\arguments{
\item{type}{Possible options are can_read / can_write / can_manage.}

\item{object}{A uuid of a project / file.}

\item{user}{A uuid of a person that permissions are taken from.}
}
\value{
Refused permissions.
}
\description{
projects_permission_refuse is a method defined in Arvados class that unables sharing files with another users_
}
