% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_permission_update}
\alias{projects_permission_update}
\title{projects_permission_update}
\usage{
arv$projects_permission_update(type_new, type_old, object, user)
}
\arguments{
\item{type_new}{New option like can_read / can_write / can_manage.}

\item{type_old}{Old option}

\item{uuid_file}{A uuid of a project / file.}

\item{uuid_person}{A uuid of the person that the permission is being updated.}
}
\value{
Updated permissions.
}
\description{
projects_permission_update is a method defined in Arvados class that enables updating permissions.
}
