% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_properties_append}
\alias{projects_properties_append}
\title{projects_properties_append}
\usage{
arv$projects_properties_append(listOfNewProperties, projectUUID)
}
\arguments{
\item{listOfNewProperties}{List of new properties.}

\item{projectUUID}{A uuid of a project / file.}
}
\value{
Project object.
}
\description{
projects_properties_append is a method defined in Arvados class that enables appending properties.
}
