% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_properties_delete}
\alias{projects_properties_delete}
\title{projects_properties_delete}
\usage{
arv$projects_properties_delete(oneProp, uuid)
}
\arguments{
\item{oneProp}{Property to be deleted.}

\item{uuid}{A uuid of a project / file.}
}
\value{
Project object.
}
\description{
projects_properties_delete is a method defined in Arvados class that enables appending properties.
}
