% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_properties_get}
\alias{projects_properties_get}
\title{projects_properties_get}
\usage{
arv$projects_properties_get(uuid)
}
\arguments{
\item{uuid}{A uuid of a project / file.}
}
\value{
Project object.
}
\description{
projects_properties_get is a method defined in Arvados class that enables appending properties.
}
