% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_properties_set}
\alias{projects_properties_set}
\title{projects_properties_set}
\usage{
arv$projects_properties_set(listProperties, projectUUID)
}
\arguments{
\item{listProperties}{List of new properties.}

\item{projectUUID}{A uuid of a project / file.}
}
\value{
Project object.
}
\description{
projects_properties_set is a method defined in Arvados class that enables setting properties.
}
