% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{projects_update}
\alias{projects_update}
\title{project.update}
\usage{
arv$projects_update(..., uuid)
}
\arguments{
\item{...}{Feature to be updated (name, description, properties)}

\item{uuid}{The UUID of the Group in question.}
}
\value{
Group object.
}
\description{
projects_update wrapps groups_update method by setting group_class attribute to "project".
}
