% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{readArvFile}
\alias{readArvFile}
\title{readArvFile}
\usage{
readArvFile(file, col, sep = ',', istable = NULL, fileclass = "SeqFastadna", Ncol = NULL, Nrow = NULL)
}
\arguments{
\item{file}{name of the file}

\item{col}{the collection from which the file is reeaded}

\item{sep}{used separator in reading tsv, csv file format}

\item{istable}{used in reading txt file to check if the file is table or not}

\item{fileclass}{used in reading fasta file to set file class}

\item{Ncol}{used in reading binary file to set numbers of columns in data.frame}

\item{Nrow}{used in reading binary file to set numbers of rows in data.frame size}
}
\value{
Collection object.
}
\description{
Read file content.
}
