% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{remove}
\alias{remove}
\title{remove}
\usage{
remove(fileNames)
}
\arguments{
\item{fileNames}{sth to be removed}
}
\value{
Collection object.
}
\description{
Remove one or more files from the collection.
}
