% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.create}
\alias{repositories.create}
\title{repositories.create}
\usage{
arv$repositories.create(repository,
	ensure_unique_name = "false")
}
\arguments{
\item{repository}{Repository object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Repository object.
}
\description{
repositories.create is a method defined in Arvados class.
}
