% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.delete}
\alias{repositories.delete}
\title{repositories.delete}
\usage{
arv$repositories.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Repository in question.}
}
\value{
Repository object.
}
\description{
repositories.delete is a method defined in Arvados class.
}
