% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.destroy}
\alias{repositories.destroy}
\title{repositories.destroy is a method defined in Arvados class.}
\usage{
arv$repositories.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Repository object.
}
\description{
repositories.destroy is a method defined in Arvados class.
}
