% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.get}
\alias{repositories.get}
\title{repositories.get}
\usage{
arv$repositories.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Repository in question.}
}
\value{
Repository object.
}
\description{
repositories.get is a method defined in Arvados class.
}
