% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.show}
\alias{repositories.show}
\title{repositories.show is a method defined in Arvados class.}
\usage{
arv$repositories.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Repository object.
}
\description{
repositories.show is a method defined in Arvados class.
}
