% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories.update}
\alias{repositories.update}
\title{repositories.update}
\usage{
arv$repositories.update(repository,
	uuid)
}
\arguments{
\item{repository}{Repository object.}

\item{uuid}{The UUID of the Repository in question.}
}
\value{
Repository object.
}
\description{
repositories.update is a method defined in Arvados class.
}
