% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories_create}
\alias{repositories_create}
\title{repositories_create}
\usage{
arv$repositories_create(repository,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{repository}{Repository object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
Repository object.
}
\description{
repositories_create is a method defined in Arvados class.
}
