% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories_delete}
\alias{repositories_delete}
\title{repositories_delete}
\usage{
arv$repositories_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Repository in question.}
}
\value{
Repository object.
}
\description{
repositories_delete is a method defined in Arvados class.
}
