% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories_get}
\alias{repositories_get}
\title{repositories_get}
\usage{
arv$repositories_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Repository in question.}
}
\value{
Repository object.
}
\description{
repositories_get is a method defined in Arvados class.
}
