% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories_get_all_permissions}
\alias{repositories_get_all_permissions}
\title{repositories_get_all_permissions}
\usage{
arv$repositories_get_all_permissions(NULL)
}
\value{
Repository object.
}
\description{
repositories_get_all_permissions is a method defined in Arvados class.
}
