% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{repositories_list}
\alias{repositories_list}
\title{repositories_list}
\usage{
arv$repositories_list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", cluster_id = NULL, bypass_federation = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}
}
\value{
RepositoryList object.
}
\description{
repositories_list is a method defined in Arvados class.
}
