% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{RepositoryList}
\alias{RepositoryList}
\title{RepositoryList}
\description{
Repository list
}
\section{Usage}{

\preformatted{repositoryList -> RepositoryList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#repositoryList.}
    \item{etag}{List version.}
    \item{items}{The list of Repositories.}
    \item{next_link}{A link to the next page of Repositories.}
    \item{next_page_token}{The page token for the next page of Repositories.}
    \item{selfLink}{A link back to this list.}
  }
}

