% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Specimen}
\alias{Specimen}
\title{Specimen}
\description{
Specimen
}
\section{Usage}{

\preformatted{specimen -> Specimen$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_by_client_uuid = NULL, modified_by_user_uuid = NULL,
	modified_at = NULL, material = NULL, updated_at = NULL,
	properties = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{material}{}
    \item{updated_at}{}
    \item{properties}{}
  }
}

