% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{SpecimenList}
\alias{SpecimenList}
\title{SpecimenList}
\description{
Specimen list
}
\section{Usage}{

\preformatted{specimenList -> SpecimenList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#specimenList.}
    \item{etag}{List version.}
    \item{items}{The list of Specimens.}
    \item{next_link}{A link to the next page of Specimens.}
    \item{next_page_token}{The page token for the next page of Specimens.}
    \item{selfLink}{A link back to this list.}
  }
}

