% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.create}
\alias{specimens.create}
\title{specimens.create}
\usage{
arv$specimens.create(specimen,
	ensure_unique_name = "false")
}
\arguments{
\item{specimen}{Specimen object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Specimen object.
}
\description{
specimens.create is a method defined in Arvados class.
}
