% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.delete}
\alias{specimens.delete}
\title{specimens.delete}
\usage{
arv$specimens.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Specimen in question.}
}
\value{
Specimen object.
}
\description{
specimens.delete is a method defined in Arvados class.
}
