% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.destroy}
\alias{specimens.destroy}
\title{specimens.destroy is a method defined in Arvados class.}
\usage{
arv$specimens.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Specimen object.
}
\description{
specimens.destroy is a method defined in Arvados class.
}
