% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.get}
\alias{specimens.get}
\title{specimens.get}
\usage{
arv$specimens.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Specimen in question.}
}
\value{
Specimen object.
}
\description{
specimens.get is a method defined in Arvados class.
}
