% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.show}
\alias{specimens.show}
\title{specimens.show is a method defined in Arvados class.}
\usage{
arv$specimens.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Specimen object.
}
\description{
specimens.show is a method defined in Arvados class.
}
