% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{specimens.update}
\alias{specimens.update}
\title{specimens.update}
\usage{
arv$specimens.update(specimen,
	uuid)
}
\arguments{
\item{specimen}{Specimen object.}

\item{uuid}{The UUID of the Specimen in question.}
}
\value{
Specimen object.
}
\description{
specimens.update is a method defined in Arvados class.
}
