% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subcollection.R
\docType{class}
\name{Subcollection-class}
\alias{Subcollection-class}
\alias{Subcollection}
\title{Subcollection Class}
\description{
Subcollection Class
}
\details{
Todo: Update description
Subcollection
}

