% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subcollection.R
\docType{data}
\name{Subcollection}
\alias{Subcollection}
\title{Arvados SubCollection Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Subcollection
}
\description{
Update description
}
\keyword{datasets}
