% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{TraitList}
\alias{TraitList}
\title{TraitList}
\description{
Trait list
}
\section{Usage}{

\preformatted{traitList -> TraitList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#traitList.}
    \item{etag}{List version.}
    \item{items}{The list of Traits.}
    \item{next_link}{A link to the next page of Traits.}
    \item{next_page_token}{The page token for the next page of Traits.}
    \item{selfLink}{A link back to this list.}
  }
}

