% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.create}
\alias{traits.create}
\title{traits.create}
\usage{
arv$traits.create(trait, ensure_unique_name = "false")
}
\arguments{
\item{trait}{Trait object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Trait object.
}
\description{
traits.create is a method defined in Arvados class.
}
