% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.delete}
\alias{traits.delete}
\title{traits.delete}
\usage{
arv$traits.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Trait in question.}
}
\value{
Trait object.
}
\description{
traits.delete is a method defined in Arvados class.
}
