% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.destroy}
\alias{traits.destroy}
\title{traits.destroy is a method defined in Arvados class.}
\usage{
arv$traits.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Trait object.
}
\description{
traits.destroy is a method defined in Arvados class.
}
