% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.get}
\alias{traits.get}
\title{traits.get}
\usage{
arv$traits.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Trait in question.}
}
\value{
Trait object.
}
\description{
traits.get is a method defined in Arvados class.
}
