% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.list}
\alias{traits.list}
\title{traits.list}
\usage{
arv$traits.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\value{
TraitList object.
}
\description{
traits.list is a method defined in Arvados class.
}
