% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.show}
\alias{traits.show}
\title{traits.show is a method defined in Arvados class.}
\usage{
arv$traits.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Trait object.
}
\description{
traits.show is a method defined in Arvados class.
}
