% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{traits.update}
\alias{traits.update}
\title{traits.update}
\usage{
arv$traits.update(trait, uuid)
}
\arguments{
\item{trait}{Trait object.}

\item{uuid}{The UUID of the Trait in question.}
}
\value{
Trait object.
}
\description{
traits.update is a method defined in Arvados class.
}
