% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{User}
\alias{User}
\title{User}
\description{
User
}
\section{Usage}{

\preformatted{user -> User$new(uuid = NULL, etag = NULL,
	owner_uuid = NULL, created_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, modified_at = NULL, email = NULL,
	first_name = NULL, last_name = NULL, identity_url = NULL,
	is_admin = NULL, prefs = NULL, updated_at = NULL, default_owner_uuid = NULL,
	is_active = NULL, username = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{email}{}
    \item{first_name}{}
    \item{last_name}{}
    \item{identity_url}{}
    \item{is_admin}{}
    \item{prefs}{}
    \item{updated_at}{}
    \item{default_owner_uuid}{}
    \item{is_active}{}
    \item{username}{}
  }
}

