% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements.create}
\alias{user_agreements.create}
\title{user_agreements.create}
\usage{
arv$user_agreements.create(useragreement,
	ensure_unique_name = "false")
}
\arguments{
\item{userAgreement}{UserAgreement object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
UserAgreement object.
}
\description{
user_agreements.create is a method defined in Arvados class.
}
