% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements.delete}
\alias{user_agreements.delete}
\title{user_agreements.delete}
\usage{
arv$user_agreements.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the UserAgreement in question.}
}
\value{
UserAgreement object.
}
\description{
user_agreements.delete is a method defined in Arvados class.
}
