% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements.destroy}
\alias{user_agreements.destroy}
\title{user_agreements.destroy is a method defined in Arvados class.}
\usage{
arv$user_agreements.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
UserAgreement object.
}
\description{
user_agreements.destroy is a method defined in Arvados class.
}
