% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements.get}
\alias{user_agreements.get}
\title{user_agreements.get}
\usage{
arv$user_agreements.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the UserAgreement in question.}
}
\value{
UserAgreement object.
}
\description{
user_agreements.get is a method defined in Arvados class.
}
