% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements.list}
\alias{user_agreements.list}
\title{user_agreements.list}
\usage{
arv$user_agreements.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\value{
UserAgreementList object.
}
\description{
user_agreements.list is a method defined in Arvados class.
}
