% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements_create}
\alias{user_agreements_create}
\title{user_agreements_create}
\usage{
arv$user_agreements_create(useragreement,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{userAgreement}{UserAgreement object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
UserAgreement object.
}
\description{
user_agreements_create is a method defined in Arvados class.
}
