% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements_new}
\alias{user_agreements_new}
\title{user_agreements_new}
\usage{
arv$user_agreements_new(NULL)
}
\value{
UserAgreement object.
}
\description{
user_agreements_new is a method defined in Arvados class.
}
