% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements_sign}
\alias{user_agreements_sign}
\title{user_agreements_sign}
\usage{
arv$user_agreements_sign(NULL)
}
\value{
UserAgreement object.
}
\description{
user_agreements_sign is a method defined in Arvados class.
}
