% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{user_agreements_update}
\alias{user_agreements_update}
\title{user_agreements_update}
\usage{
arv$user_agreements_update(useragreement,
	uuid)
}
\arguments{
\item{userAgreement}{UserAgreement object.}

\item{uuid}{The UUID of the UserAgreement in question.}
}
\value{
UserAgreement object.
}
\description{
user_agreements_update is a method defined in Arvados class.
}
