% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{UserAgreement}
\alias{UserAgreement}
\title{UserAgreement}
\description{
UserAgreement
}
\section{Usage}{

\preformatted{userAgreement -> UserAgreement$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_by_client_uuid = NULL, modified_by_user_uuid = NULL,
	modified_at = NULL, portable_data_hash = NULL, replication_desired = NULL,
	replication_confirmed_at = NULL, replication_confirmed = NULL,
	updated_at = NULL, manifest_text = NULL, name = NULL,
	description = NULL, properties = NULL, delete_at = NULL,
	file_names = NULL, trash_at = NULL, is_trashed = NULL,
	storage_classes_desired = NULL, storage_classes_confirmed = NULL,
	storage_classes_confirmed_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{modified_at}{}
    \item{portable_data_hash}{}
    \item{replication_desired}{}
    \item{replication_confirmed_at}{}
    \item{replication_confirmed}{}
    \item{updated_at}{}
    \item{uuid}{}
    \item{manifest_text}{}
    \item{name}{}
    \item{description}{}
    \item{properties}{}
    \item{delete_at}{}
    \item{file_names}{}
    \item{trash_at}{}
    \item{is_trashed}{}
    \item{storage_classes_desired}{}
    \item{storage_classes_confirmed}{}
    \item{storage_classes_confirmed_at}{}
  }
}

