% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{UserAgreementList}
\alias{UserAgreementList}
\title{UserAgreementList}
\description{
UserAgreement list
}
\section{Usage}{

\preformatted{userAgreementList -> UserAgreementList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#userAgreementList.}
    \item{etag}{List version.}
    \item{items}{The list of UserAgreements.}
    \item{next_link}{A link to the next page of UserAgreements.}
    \item{next_page_token}{The page token for the next page of UserAgreements.}
    \item{selfLink}{A link back to this list.}
  }
}

