% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{UserList}
\alias{UserList}
\title{UserList}
\description{
User list
}
\section{Usage}{

\preformatted{userList -> UserList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#userList.}
    \item{etag}{List version.}
    \item{items}{The list of Users.}
    \item{next_link}{A link to the next page of Users.}
    \item{next_page_token}{The page token for the next page of Users.}
    \item{selfLink}{A link back to this list.}
  }
}

