% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.activate}
\alias{users.activate}
\title{users.activate}
\usage{
arv$users.activate(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
User object.
}
\description{
users.activate is a method defined in Arvados class.
}
