% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.create}
\alias{users.create}
\title{users.create}
\usage{
arv$users.create(user, ensure_unique_name = "false")
}
\arguments{
\item{user}{User object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
User object.
}
\description{
users.create is a method defined in Arvados class.
}
