% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.delete}
\alias{users.delete}
\title{users.delete}
\usage{
arv$users.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users.delete is a method defined in Arvados class.
}
