% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.destroy}
\alias{users.destroy}
\title{users.destroy is a method defined in Arvados class.}
\usage{
arv$users.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
User object.
}
\description{
users.destroy is a method defined in Arvados class.
}
