% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.get}
\alias{users.get}
\title{users.get is a method defined in Arvados class.}
\usage{
arv$users.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users.get is a method defined in Arvados class.
}
