% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.setup}
\alias{users.setup}
\title{users.setup}
\usage{
arv$users.setup(user = NULL, openid_prefix = NULL,
	repo_name = NULL, vm_uuid = NULL, send_notification_email = "false")
}
\arguments{
\item{user}{}

\item{openid_prefix}{}

\item{repo_name}{}

\item{vm_uuid}{}

\item{send_notification_email}{}
}
\value{
User object.
}
\description{
users.setup is a method defined in Arvados class.
}
