% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.show}
\alias{users.show}
\title{users.show is a method defined in Arvados class.}
\usage{
arv$users.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
User object.
}
\description{
users.show is a method defined in Arvados class.
}
