% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.unsetup}
\alias{users.unsetup}
\title{users.unsetup}
\usage{
arv$users.unsetup(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
User object.
}
\description{
users.unsetup is a method defined in Arvados class.
}
