% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.update}
\alias{users.update}
\title{users.update}
\usage{
arv$users.update(user, uuid)
}
\arguments{
\item{user}{User object.}

\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users.update is a method defined in Arvados class.
}
