% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.update_uuid}
\alias{users.update_uuid}
\title{users.update_uuid}
\usage{
arv$users.update_uuid(uuid, new_uuid)
}
\arguments{
\item{uuid}{}

\item{new_uuid}{}
}
\value{
User object.
}
\description{
users.update_uuid is a method defined in Arvados class.
}
