% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_create}
\alias{users_create}
\title{users_create}
\usage{
arv$users_create(user, ensure_unique_name = "false",
	cluster_id = NULL)
}
\arguments{
\item{user}{User object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
User object.
}
\description{
users_create is a method defined in Arvados class.
}
