% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_delete}
\alias{users_delete}
\title{users_delete}
\usage{
arv$users_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users_delete is a method defined in Arvados class.
}
