% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_get}
\alias{users_get}
\title{users_get}
\usage{
arv$users_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users_get is a method defined in Arvados class.
}
