% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_merge}
\alias{users_merge}
\title{users_merge}
\usage{
arv$users_merge(new_owner_uuid,
	new_user_token = NULL, redirect_to_new_user = NULL,
	old_user_uuid = NULL, new_user_uuid = NULL)
}
\value{
User object.
}
\description{
users_merge is a method defined in Arvados class.
}
