% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_setup}
\alias{users_setup}
\title{users_setup}
\usage{
arv$users_setup(uuid = NULL, user = NULL,
	repo_name = NULL, vm_uuid = NULL, send_notification_email = "false")
}
\value{
User object.
}
\description{
users_setup is a method defined in Arvados class.
}
