% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users_update}
\alias{users_update}
\title{users_update}
\usage{
arv$users_update(user, uuid, bypass_federation = NULL)
}
\arguments{
\item{user}{User object.}

\item{uuid}{The UUID of the User in question.}
}
\value{
User object.
}
\description{
users_update is a method defined in Arvados class.
}
