% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.create}
\alias{virtual_machines.create}
\title{virtual_machines.create}
\usage{
arv$virtual_machines.create(virtualmachine,
	ensure_unique_name = "false")
}
\arguments{
\item{virtualMachine}{VirtualMachine object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.create is a method defined in Arvados class.
}
