% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.get}
\alias{virtual_machines.get}
\title{virtual_machines.get}
\usage{
arv$virtual_machines.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the VirtualMachine in question.}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.get is a method defined in Arvados class.
}
