% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.get_all_logins}
\alias{virtual_machines.get_all_logins}
\title{virtual_machines.get_all_logins}
\usage{
arv$virtual_machines.get_all_logins(NULL)
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.get_all_logins is a method defined in Arvados class.
}
