% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.list}
\alias{virtual_machines.list}
\title{virtual_machines.list}
\usage{
arv$virtual_machines.list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact")
}
\arguments{
\item{filters}{}

\item{where}{}

\item{order}{}

\item{select}{}

\item{distinct}{}

\item{limit}{}

\item{offset}{}

\item{count}{}
}
\value{
VirtualMachineList object.
}
\description{
virtual_machines.list is a method defined in Arvados class.
}
