% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.logins}
\alias{virtual_machines.logins}
\title{virtual_machines.logins}
\usage{
arv$virtual_machines.logins(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.logins is a method defined in Arvados class.
}
