% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.show}
\alias{virtual_machines.show}
\title{virtual_machines.show is a method defined in Arvados class.}
\usage{
arv$virtual_machines.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.show is a method defined in Arvados class.
}
