% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines.update}
\alias{virtual_machines.update}
\title{virtual_machines.update}
\usage{
arv$virtual_machines.update(virtualmachine,
	uuid)
}
\arguments{
\item{virtualMachine}{VirtualMachine object.}

\item{uuid}{The UUID of the VirtualMachine in question.}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines.update is a method defined in Arvados class.
}
