% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines_create}
\alias{virtual_machines_create}
\title{virtual_machines_create}
\usage{
arv$virtual_machines_create(virtualmachine,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{virtualMachine}{VirtualMachine object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines_create is a method defined in Arvados class.
}
