% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines_delete}
\alias{virtual_machines_delete}
\title{virtual_machines_delete}
\usage{
arv$virtual_machines_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the VirtualMachine in question.}
}
\value{
VirtualMachine object.
}
\description{
virtual_machines_delete is a method defined in Arvados class.
}
