% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines_list}
\alias{virtual_machines_list}
\title{virtual_machines_list}
\usage{
arv$virtual_machines_list(filters = NULL,
	where = NULL, order = NULL, select = NULL,
	distinct = NULL, limit = "100", offset = "0",
	count = "exact", cluster_id = NULL, bypass_federation = NULL)
}
\arguments{
\item{cluster_id}{List objects on a remote federated cluster instead of the current one.}

\item{bypass_federation}{bypass federation behavior, list items from local instance database only}
}
\value{
VirtualMachineList object.
}
\description{
virtual_machines_list is a method defined in Arvados class.
}
