% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{virtual_machines_logins}
\alias{virtual_machines_logins}
\title{virtual_machines_logins}
\usage{
arv$virtual_machines_logins(uuid)
}
\value{
VirtualMachine object.
}
\description{
virtual_machines_logins is a method defined in Arvados class.
}
