% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{VirtualMachineList}
\alias{VirtualMachineList}
\title{VirtualMachineList}
\description{
VirtualMachine list
}
\section{Usage}{

\preformatted{virtualMachineList -> VirtualMachineList$new(kind = NULL,
	etag = NULL, items = NULL, next_link = NULL, next_page_token = NULL,
	selfLink = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{kind}{Object type. Always arvados#virtualMachineList.}
    \item{etag}{List version.}
    \item{items}{The list of VirtualMachines.}
    \item{next_link}{A link to the next page of VirtualMachines.}
    \item{next_page_token}{The page token for the next page of VirtualMachines.}
    \item{selfLink}{A link back to this list.}
  }
}

