% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArvadosClasses.R
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\description{
Workflow
}
\section{Usage}{

\preformatted{workflow -> Workflow$new(uuid = NULL,
	etag = NULL, owner_uuid = NULL, created_at = NULL,
	modified_at = NULL, modified_by_client_uuid = NULL,
	modified_by_user_uuid = NULL, name = NULL, description = NULL,
	definition = NULL, updated_at = NULL)
}
}

\section{Arguments}{

  \describe{
    \item{uuid}{Object ID.}
    \item{etag}{Object version.}
    \item{uuid}{}
    \item{owner_uuid}{}
    \item{created_at}{}
    \item{modified_at}{}
    \item{modified_by_client_uuid}{}
    \item{modified_by_user_uuid}{}
    \item{name}{}
    \item{description}{}
    \item{definition}{}
    \item{updated_at}{}
  }
}

