% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows.create}
\alias{workflows.create}
\title{workflows.create}
\usage{
arv$workflows.create(workflow,
	ensure_unique_name = "false")
}
\arguments{
\item{workflow}{Workflow object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision.}
}
\value{
Workflow object.
}
\description{
workflows.create is a method defined in Arvados class.
}
