% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows.delete}
\alias{workflows.delete}
\title{workflows.delete}
\usage{
arv$workflows.delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Workflow in question.}
}
\value{
Workflow object.
}
\description{
workflows.delete is a method defined in Arvados class.
}
