% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows.destroy}
\alias{workflows.destroy}
\title{workflows.destroy is a method defined in Arvados class.}
\usage{
arv$workflows.destroy(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Workflow object.
}
\description{
workflows.destroy is a method defined in Arvados class.
}
