% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows.get}
\alias{workflows.get}
\title{workflows.get}
\usage{
arv$workflows.get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Workflow in question.}
}
\value{
Workflow object.
}
\description{
workflows.get is a method defined in Arvados class.
}
