% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows.show}
\alias{workflows.show}
\title{workflows.show is a method defined in Arvados class.}
\usage{
arv$workflows.show(uuid)
}
\arguments{
\item{uuid}{}
}
\value{
Workflow object.
}
\description{
workflows.show is a method defined in Arvados class.
}
