% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows_create}
\alias{workflows_create}
\title{workflows_create}
\usage{
arv$workflows_create(workflow,
	ensure_unique_name = "false", cluster_id = NULL)
}
\arguments{
\item{workflow}{Workflow object.}

\item{ensure_unique_name}{Adjust name to ensure uniqueness instead of returning an error on (owner_uuid, name) collision_}

\item{cluster_id}{Create object on a remote federated cluster instead of the current one.}
}
\value{
Workflow object.
}
\description{
workflows_create is a method defined in Arvados class.
}
