% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows_delete}
\alias{workflows_delete}
\title{workflows_delete}
\usage{
arv$workflows_delete(uuid)
}
\arguments{
\item{uuid}{The UUID of the Workflow in question.}
}
\value{
Workflow object.
}
\description{
workflows_delete is a method defined in Arvados class.
}
