% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows_get}
\alias{workflows_get}
\title{workflows_get}
\usage{
arv$workflows_get(uuid)
}
\arguments{
\item{uuid}{The UUID of the Workflow in question.}
}
\value{
Workflow object.
}
\description{
workflows_get is a method defined in Arvados class.
}
