% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{workflows_update}
\alias{workflows_update}
\title{workflows_update}
\usage{
arv$workflows_update(workflow,
	uuid)
}
\arguments{
\item{workflow}{Workflow object.}

\item{uuid}{The UUID of the Workflow in question.}
}
\value{
Workflow object.
}
\description{
workflows_update is a method defined in Arvados class.
}
