% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\docType{data}
\name{Arvados}
\alias{Arvados}
\title{Arvados SDK Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Arvados
}
\description{
All Arvados logic is inside this class
}
\section{Fields}{

\describe{
\item{\code{token}}{Token represents user authentification token.}

\item{\code{host}}{Host represents server name we wish to connect to.}
}}

\examples{
arv = Arvados$new("token", "host_name")
}
\keyword{datasets}
