% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{authorized_keys.update}
\alias{authorized_keys.update}
\title{authorized_keys.update}
\usage{
arv$authorized_keys.update(authorizedkey,
	uuid)
}
\arguments{
\item{authorizedKey}{AuthorizedKey object.}

\item{uuid}{The UUID of the AuthorizedKey in question.}
}
\value{
AuthorizedKey object.
}
\description{
authorized_keys.update is a method defined in Arvados class.
}
