% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arvados_objects.R
\name{Collection}
\alias{Collection}
\title{Collection Object}
\usage{
Collection(uuid = NULL, etag = NULL, owner_uuid = NULL,
  created_at = NULL, modified_by_client_uuid = NULL,
  modified_by_user_uuid = NULL, modified_at = NULL,
  portable_data_hash = NULL, replication_desired = NULL,
  replication_confirmed_at = NULL, replication_confirmed = NULL,
  updated_at = NULL, manifest_text = NULL, name = NULL,
  description = NULL, properties = NULL, delete_at = NULL,
  file_names = NULL, trash_at = NULL, is_trashed = NULL)
}
\arguments{
\item{uuid}{Object ID}

\item{etag}{Object version}

\item{owner_uuid}{No description}

\item{created_at}{No description}

\item{modified_by_client_uuid}{No description}

\item{modified_by_user_uuid}{No description}

\item{modified_at}{No description}

\item{portable_data_hash}{No description}

\item{replication_desired}{No description}

\item{replication_confirmed_at}{No description}

\item{replication_confirmed}{No description}

\item{updated_at}{No description}

\item{manifest_text}{No description}

\item{name}{No description}

\item{description}{No description}

\item{properties}{No description}

\item{delete_at}{No description}

\item{file_names}{No description}

\item{trash_at}{No description}

\item{is_trashed}{No description}
}
\value{
Collection object
}
\description{
Collection Object
}
\details{
Todo: Update description
Collection
}
\concept{Collection functions}
