% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{Collection}
\alias{Collection}
\title{Collection}
\description{
Collection class provides interface for working with Arvados collections.
}
\section{Usage}{

\preformatted{collection = Collection$new(arv, uuid)}
}

\section{Arguments}{

\describe{
  \item{arv}{Arvados object.}
  \item{uuid}{UUID of a collection.}
}
}

\section{Methods}{

\describe{
  \item{add(content)}{Adds ArvadosFile or Subcollection specified by content to the collection.}
  \item{create(fileNames, relativePath = "")}{Creates one or more ArvadosFiles and adds them to the collection at specified path.}
  \item{remove(fileNames)}{Remove one or more files from the collection.}
  \item{move(content, newLocation)}{Moves ArvadosFile or Subcollection to another location in the collection.}
  \item{getFileListing()}{Returns collections file content as character vector.}
  \item{get(relativePath)}{If relativePath is valid, returns ArvadosFile or Subcollection specified by relativePath, else returns NULL.}
}
}

\examples{
\dontrun{
arv <- Arvados$new("your Arvados token", "example.arvadosapi.com")
collection <- Collection$new(arv, "uuid")

newFile <- ArvadosFile$new("myFile")
collection$add(newFile, "myFolder")

createdFiles <- collection$create(c("main.cpp", lib.dll), "cpp/src/")

collection$remove("location/to/my/file.cpp")

collection$move("folder/file.cpp", "file.cpp")

arvadosFile <- collection$get("location/to/my/file.cpp")
arvadosSubcollection <- collection$get("location/to/my/directory/")
}
}
