% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{Collection}
\alias{Collection}
\title{Collection}
\description{
Collection class provides interface for working with Arvados collections,
for exaplme actions like creating, updating, moving or removing are possible.
}
\section{Usage}{

\preformatted{collection = Collection$new(arv, uuid)}
}

\section{Arguments}{

\describe{
  \item{arv}{Arvados object.}
  \item{uuid}{UUID of a collection.}
}
}

\section{Methods}{

\describe{
	\item{}{\code{\link{add}}}
	\item{}{\code{\link{create}}}
	\item{}{\code{\link{remove}}}
	\item{}{\code{\link{move}}}
	\item{}{\code{\link{readArvFile}}}
	\item{}{\code{\link{copy}}}
	\item{}{\code{\link{getFileListing}}}
	\item{}{\code{\link{get}}}
}
}

\examples{
\dontrun{
# initialize API and a collection object:
arv <- Arvados$new("your Arvados token", "example.arvadosapi.com")
collection <- Collection$new(arv, "uuid")

# Create new file in a collection
mainFile <- collection$create("cpp/src/main.cpp")[[1]]
fileList <- collection$create(c("main.cpp", lib.dll), "cpp/src/")
# NOTE: It returns a vector of one or more ArvadosFile objects

# Get ArvadosFile or Subcollection from internal tree-like structure
arvadosFile <- collection$get("location/to/my/file.cpp")
arvadosSubcollection <- collection$get("location/to/my/directory/")

# Delete file or list of them from a collection
collection$remove("location/to/my/file.cpp") # delete file
collection$remove(c("path/to/my/file.cpp", "path/to/other/file.cpp")) # delete list of files
# NOTE: You can remove both Subcollection and ArvadosFile. If subcollection contains more files or folders they will be removed recursively.

# Move or rename a file or folder within a collection
collection$move("folder/file.cpp", "file.cpp")
}
}
\seealso{
\code{\link{https://github.com/arvados/arvados/tree/main/sdk/R}}
}
