% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subcollection.R
\name{Subcollection}
\alias{Subcollection}
\title{Subcollection}
\description{
Subcollection class represents a folder inside Arvados collection.
It is essentially a composite of arvadosFiles and other subcollections.
}
\section{Usage}{

\preformatted{subcollection = Subcollection$new(name)}
}

\section{Arguments}{

\describe{
  \item{name}{Name of the subcollection.}
}
}

\section{Methods}{

\describe{
  \item{}{\code{\link{getName}}}
  \item{}{\code{\link{getRelativePath}}}
  \item{}{\code{\link{add}}}
  \item{}{\code{\link{remove}}}
  \item{}{\code{\link{get}}}
  \item{}{\code{\link{getFileListing}}}
  \item{}{\code{\link{getSizeInBytes}}}
  \item{}{\code{\link{move}}}
  \item{}{\code{\link{copy}}}
}
}

\examples{
\dontrun{
myFolder <- Subcollection$new("myFolder")
myFile   <- ArvadosFile$new("myFile")

myFolder$add(myFile)
myFolder$get("myFile")
myFolder$remove("myFile")

myFolder$move("newLocation/myFolder")
myFolder$copy("newLocation/myFolder")
}
}
