% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arvados.R
\name{users.merge}
\alias{users.merge}
\title{users.merge}
\usage{
arv$users.merge(new_owner_uuid,
	new_user_token, redirect_to_new_user = NULL)
}
\arguments{
\item{new_owner_uuid}{}

\item{new_user_token}{}

\item{redirect_to_new_user}{}
}
\value{
User object.
}
\description{
users.merge is a method defined in Arvados class.
}
